Major refactor, replacing one of the two openssl commands.
This is a total refactor of analyzing a cert chain for SHA-ness.
It now downloads the full certificate chain and runs the check on
each certificate in the chain.

Instead of one big giant shell command to get some text for the
signature algorithm, this now has `openssl s_client` fetch the full
certificate chain using -showcerts. Then, the resulting cert chain is
parsed in JavaScript to grab each cert.

Each cert is then run through a Node library, x509, to get the
signature algorithm field. We take this opportunity to then fetch a
few more useful fields from the x509 metadata, like expiration date
and subject name.

Tests pass, but this very well may cough up unexpected errors on
the wild world of certs out there. Then again, the last version was
relying on a `sed` regex, and this regex can't be much worse than
that.